Track depkeys that are initially undefined
When the userland schema invokes `schemaInterface.getAttr(name)` we
track a dependency between `name` and the key being resolved.

Change that tracking to do so even when the value within `modelData` is
`undefined`; ie track keys that are requested, not only those that have
values.

This is generally not needed, but is necessary if the schema relies on
property relationships whose keys can change between requests.
