+++ tests/unit/model-test.js
@@ -1023,112 +1023,6 @@ module('unit/model', function(hooks) {
     );
   });
 
-  test('upon updating the data in store, attributes refering to keys starting with `*` should update', function(assert) {
-    let model = run(() => {
-      return this.store.push({
-        data: {
-          id: 'isbn:9780439708180',
-          type: 'com.example.bookstore.Book',
-          attributes: {
-            name: `Harry Potter and the Sorcerer's Stone`,
-            pubDate: 'September 1989',
-            '*relatedBooks': ['isbn:9780439358079'],
-            '*otherBooksInSeries': ['isbn:9780439358071', 'isbn:9780439136365'],
-          },
-        },
-        included: [
-          {
-            id: 'isbn:9780439358071',
-            type: 'com.example.bookstore.Book',
-            attributes: {
-              name: `Harry Potter and the Chamber of Secrets`,
-            },
-          },
-          {
-            id: 'isbn:9780439136365',
-            type: 'com.example.bookstore.Book',
-            attributes: {
-              name: `Harry Potter and the Prisoner of Azkaban`,
-            },
-          },
-          {
-            id: 'isbn:9780439358079',
-            type: 'com.example.bookstore.Book',
-            attributes: {
-              name: `Fantastic Beasts and Where to Find Them`,
-            },
-          },
-        ],
-      });
-    });
-
-    let otherBooks = get(model, 'otherBooksInSeries');
-    let relatedBooks = get(model, 'relatedBooks');
-    assert.deepEqual(
-      otherBooks.map(b => get(b, 'name')),
-      ['Harry Potter and the Chamber of Secrets', 'Harry Potter and the Prisoner of Azkaban'],
-      'attr array ref is array-like'
-    );
-    assert.deepEqual(
-      relatedBooks.map(b => get(b, 'name')),
-      ['Fantastic Beasts and Where to Find Them'],
-      'attr array ref is array-like'
-    );
-
-    //Update record with new data
-    run(() =>
-      this.store.push({
-        data: {
-          id: 'isbn:9780439708180',
-          type: 'com.example.bookstore.Book',
-          attributes: {
-            name: `Harry Potter and the Sorcerer's Stone`,
-            pubDate: 'September 1989',
-            '*relatedBooks': ['isbn:9780439358080'],
-            '*otherBooksInSeries': ['isbn:9780439064878', 'isbn:9780439064879'],
-          },
-        },
-        included: [
-          {
-            id: 'isbn:9780439064878',
-            type: 'com.example.bookstore.Book',
-            attributes: {
-              name: `Harry Potter and the Goblet of Fire`,
-            },
-          },
-          {
-            id: 'isbn:9780439064879',
-            type: 'com.example.bookstore.Book',
-            attributes: {
-              name: `Harry Potter and the Order of the Phoenix`,
-            },
-          },
-          {
-            id: 'isbn:9780439358080',
-            type: 'com.example.bookstore.Book',
-            attributes: {
-              name: `Fantastic Beasts and Where to Find Them 2`,
-            },
-          },
-        ],
-      })
-    );
-
-    model = this.store.peekRecord('com.example.bookstore.Book', 'isbn:9780439708180');
-    otherBooks = get(model, 'otherBooksInSeries');
-    relatedBooks = get(model, 'relatedBooks');
-    assert.deepEqual(
-      otherBooks.map(b => get(b, 'name')),
-      ['Harry Potter and the Goblet of Fire', 'Harry Potter and the Order of the Phoenix'],
-      'attr ref is updated upon reload'
-    );
-    assert.deepEqual(
-      relatedBooks.map(b => get(b, 'name')),
-      ['Fantastic Beasts and Where to Find Them 2'],
-      'attr array is updated upon reload'
-    );
-  });
-
   test('default values are not transformed', function(assert) {
     let model = run(() =>
       this.store.push({
